Test node is manager in nonWorkerManager cluster
This testcase is to check whether the node is treated as a manager or not when the clustering pattern is set to nonWorkerManager.
The correct behavior is to consider this node as neither worker nor manager as the clustering pattern is nonWorkerManager. This can be tracked by the log message in the carbon log. If the node is treated as manager node, it prevents running the inbound endpoint in manager node and prints "InboundRunner Inbound EP will not run in manager node. Same will run on worker(s)" as an INFO log
The associated issue is : https://github.com/wso2/product-ei/issues/532