Fix test failure in RestPostFixUrlTest
java.net.SocketTimeoutException: Read timed out exception was thrown when invoke an API. This can be happened the same API was exposed via another test which has run parallely. (This is an intermittent issue.).
Therefore, as a solution the API url was renamed to be unique.
Further, in order to avoid intermittent issues related to reading logs, the Utils.checkForLog() method is used.