Fix validate mediator test failure (#1754)
Previously, when the source element is given and if it is not accessible from the request body, logic was written to throw a synapse exception and handle it inside the validate mediator by forwarding to fault sequence of the validate mediator. Now this fix catches that synapse exception and invokes the on-fail sequence since it is a validation error. Test case has to be altered to follow the progressions Git Issue: https://github.com/wso2/product-ei/issues/1753