Improve startup time of the micro-integrator (#2426)
* Modify distribution directory name of the profile-creator.sh

Added a parameter to profile-creator bash script to get the current version of the distribution for the distribution directory name

* Parameterize extension loading in micro-integrator

Added a system property which can be defined in the wso2server.sh in order to enable/disable extension loading. The purpose of this modification is to enhance the startup time of the server when extensions are not in use.